try moving frontend's dbconn to shared package
