+++ cmd/frontend/db/org_invitations.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/keegancsmith/sqlf"
 	"github.com/lib/pq"
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/db/dbconn"
+	"github.com/sourcegraph/sourcegraph/pkg/dbconn"
 )
 
 // An OrgInvitation is an invitation for a user to join an organization as a member.
