+++ enterprise/cmd/frontend/internal/dotcom/productsubscription/licenses_db.go
@@ -8,7 +8,7 @@ import (
 	"github.com/keegancsmith/sqlf"
 	"github.com/pkg/errors"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/db"
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/db/dbconn"
+	"github.com/sourcegraph/sourcegraph/pkg/dbconn"
 )
 
 // dbLicense describes an product license row in the product_licenses DB table.
