+++ enterprise/cmd/frontend/internal/registry/releases_db.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/keegancsmith/sqlf"
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/db/dbconn"
+	"github.com/sourcegraph/sourcegraph/pkg/dbconn"
 )
 
 // dbRelease describes a release of an extension in the extension registry.
